collections: dotdict() moved to collections.py
