+++ neurolib/models/model.py
@@ -2,20 +2,7 @@ import logging
 import xarray as xr
 import numpy as np
 
-
-class dotdict(dict):
-    """dot.notation access to dictionary attributes"""
-
-    __getattr__ = dict.get
-    __setattr__ = dict.__setitem__
-    __delattr__ = dict.__delitem__
-
-    # now pickleable!!!
-    def __getstate__(self):
-        return dict(self)
-
-    def __setstate__(self, state):
-        self.update(state)
+from neurolib.utils.collections import dotdict
 
 
 class Model:
